Set up data store for login status
