+++ app/src/main/java/com/binar/gosky/data/local/preference/UserDataStoreManager.kt
@@ -1,0 +1,32 @@
+package com.binar.gosky.data.local.preference
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.*
+import androidx.datastore.preferences.preferencesDataStore
+import dagger.hilt.android.qualifiers.ApplicationContext
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+import javax.inject.Inject
+
+class UserDataStoreManager @Inject constructor(@ApplicationContext private val context: Context) {
+
+    suspend fun setUserLogin(isLogin: Boolean) {
+        context.userDataStore.edit { preferences ->
+            preferences[LOGIN_STATUS_KEY] = isLogin
+        }
+    }
+
+    fun getUserLogin(): Flow<Boolean> {
+        return context.userDataStore.data.map { preferences ->
+            preferences[LOGIN_STATUS_KEY] ?: false
+        }
+    }
+
+    companion object {
+        private const val DATA_STORE_NAME = "user_preferences"
+        private val LOGIN_STATUS_KEY = booleanPreferencesKey("login_status_key")
+
+        val Context.userDataStore: DataStore<Preferences> by preferencesDataStore(name = DATA_STORE_NAME)
+    }
+}
