+++ app/src/main/java/com/binar/gosky/presentation/ui/auth/EmailConfirmationFragment.kt
@@ -1,4 +1,4 @@
-package com.binar.gosky.presentation.ui.home
+package com.binar.gosky.presentation.ui.auth
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,14 +7,14 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.binar.gosky.R
 
-class EmailVerificationNotificationFragment : Fragment() {
+class EmailConfirmationFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_email_verification_notification, container, false)
+        return inflater.inflate(R.layout.fragment_email_confirmation, container, false)
     }
 
 }
