+++ app/src/main/java/com/binar/gosky/presentation/ui/auth/EmailVerificationNotificationFragment.kt
@@ -1,20 +1,20 @@
-package com.binar.gosky.presentation.ui.home
+package com.binar.gosky.presentation.ui.auth
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.binar.gosky.R
 
-class RegisterFragment : Fragment() {
+class EmailVerificationNotificationFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_register, container, false)
+        return inflater.inflate(R.layout.fragment_email_verification_notification, container, false)
     }
 
 }
