+++ app/src/main/res/layout/fragment_email_confirmation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue_background"
-    tools:context=".presentation.ui.home.EmailConfirmationFragment">
+    tools:context=".presentation.ui.auth.EmailConfirmationFragment">
 
 
     <TextView
