+++ app/src/main/res/layout/fragment_email_verification_notification.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue_background"
-    tools:context=".presentation.ui.home.EmailVerificationNotificationFragment">
+    tools:context=".presentation.ui.auth.EmailVerificationNotificationFragment">
 
 
     <ImageView
