+++ app/src/main/res/layout/fragment_forgot_password.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.home.ForgotPasswordFragment">
+    tools:context=".presentation.ui.auth.ForgotPasswordFragment">
 
 
     <com.google.android.material.appbar.AppBarLayout
