+++ app/src/main/res/layout/fragment_register.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.home.RegisterFragment">
+    tools:context=".presentation.ui.auth.RegisterFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/app_bar"
